# swagger.yaml
openapi: 3.0.3
info:
  title: Elice SW Track 7th Web Project Team 5 - 온앤오프
  description: API 문서
  version: 1.0.0
servers:
  - url: http://localhost:4000/api
...
paths:
  /auth/login/{social}:
    get:
      tags:
        - auth
      summary: 소셜 로그인
      description: 소셜 로그인 API
      parameters:
        - name: social
          in: path
          description: 소셜 ID(google, kakao, facebook, naver, github 중 1)
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 성공
        '4xx':
          description: 잘못된 요청
        '5xx':
          description: 서버 에러
  /auth/login/{social}/callback:
    get:
      tags:
        - auth
      summary: 소셜 로그인(콜백)
      description: 소셜 로그인(콜백) API
      parameters:
        - name: social
          in: path
          description: 소셜 ID(google, kakao, facebook, naver, github 중 1)
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 성공
        '4xx':
          description: 잘못된 요청
        '5xx':
          description: 서버 에러
  /auth/logout:
    get:
      tags:
        - auth
      summary: 로그아웃
      description: 로그아웃 API
      responses:
        '200':
          description: 성공
        '4xx':
          description: 잘못된 요청
        '5xx':
          description: 서버 에러
  /auth/withdraw:
    get:
      tags:
        - auth
      summary: 회원탈퇴
      description: 회원탈퇴 API
      responses:
        '200':
          description: 성공
        '4xx':
          description: 잘못된 요청
        '5xx':
          description: 서버 에러
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer